refactor: eliminate 5 sets of duplicated helpers across PracticeSession coordinators
- Extract audioErrorText() and uniqueMIDINotesByHand() as shared free functions
- Extract recordPlaybackError() and strictTriggerGuideIndex() onto PracticeSessionStateStore
- Remove identical copies from PlaybackCoordinator, ManualReplayCoordinator,
  HighlightGuideController, HandGateController, and ViewModelCommands
- All call sites updated to use stateStore.* or free function equivalents
