+++ LonelyPianistAVP/ViewModels/PracticeSession/Input/PracticeHandGateController.swift
@@ -88,20 +88,5 @@ final class PracticeHandGateController: PracticeSessionLifecycleProtocol {
             effectHandler?.handle(effect: .advanceToNextStep)
         }
     }
-
-    private func uniqueMIDINotesByHand(in step: PracticeStep) -> (right: [Int], left: [Int]) {
-        var right: Set<Int> = []
-        var left: Set<Int> = []
-
-        for note in step.notes {
-            if note.hand == .left {
-                left.insert(note.midiNote)
-            } else {
-                right.insert(note.midiNote)
-            }
-        }
-
-        return (right: right.sorted(), left: left.sorted())
-    }
 }
