+++ LonelyPianistAVP/ViewModels/PracticeSession/PracticeSessionViewModelCommands.swift
@@ -444,32 +444,4 @@ extension PracticeSessionViewModel {
         Set(step.notes.map(\.midiNote)).sorted()
     }
 
-    func uniqueMIDINotesByHand(in step: PracticeStep) -> (right: [Int], left: [Int]) {
-        var right: Set<Int> = []
-        var left: Set<Int> = []
-
-        for note in step.notes {
-            if note.hand == .left {
-                left.insert(note.midiNote)
-            } else {
-                right.insert(note.midiNote)
-            }
-        }
-
-        return (right: right.sorted(), left: left.sorted())
-    }
-
-    func recordPlaybackError(_ error: Error) {
-        guard self.audioPlaybackErrorMessage == nil else { return }
-        self.audioPlaybackErrorMessage = audioErrorText(for: error)
-    }
-
-    func audioErrorText(for error: Error) -> String {
-        if let localized = error as? LocalizedError, let description = localized.errorDescription,
-           description.isEmpty == false
-        {
-            return description
-        }
-        return String(describing: error)
-    }
 }
